pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lirox"
version = "2.0.0"
description = "Lirox — Autonomous AI Agent OS with pluggable skills"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "Baljot Singh"}]
keywords = ["ai", "agent", "cli", "autonomous", "llm", "skills"]

dependencies = [
    "rich>=13.0.0",
    "prompt_toolkit>=3.0.0",
    "psutil>=5.9.0",
    "google-genai>=0.3.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "schedule>=1.2.0",
]

[project.optional-dependencies]
full = ["openai>=1.0.0", "anthropic>=0.7.0", "groq>=0.4.0", "websockets>=12.0"]

[project.scripts]
lirox = "lirox.main:main"

[tool.setuptools.packages.find]
include = ["lirox*"]
